Refactoring types into H3.Data module.
