+++ src/H3/Data.hs
@@ -1,0 +1,16 @@
+module H3.Data
+  ( LatLng(LatLng)
+  , H3Index
+  , H3ErrorCodes(..)
+  , GeoPolygon(GeoPolygon)
+  , CoordIJ(CoordIJ)
+  ) where
+
+import H3.Internal.H3Api 
+  ( LatLng(LatLng)
+  , H3ErrorCodes(..)
+  , H3Index
+  , GeoPolygon(GeoPolygon)
+  , CoordIJ(CoordIJ)
+  )
+
