+++ src/H3/Traversal.hs
@@ -1,6 +1,5 @@
 module H3.Traversal
-  ( CoordIJ(CoordIJ)
-  , gridDisk
+  ( gridDisk
   , gridDiskUnsafe
   , gridDiskDistances
   , gridDiskDistancesSafe
@@ -16,7 +15,7 @@ import Data.Int (Int64)
 import H3.Internal.H3Api 
   ( H3ErrorCodes
   , H3Index
-  , CoordIJ(CoordIJ)
+  , CoordIJ
   , c2hs_gridDistance
   , c2hs_cellToLocalIj
   , c2hs_localIjToCell
