+++ src/H3/Vertexes.hs
@@ -5,9 +5,9 @@ module H3.Vertexes
   , isValidVertex
   ) where
 
-import H3.Indexing (LatLng)
 import H3.Internal.H3Api 
-  ( H3ErrorCodes
+  ( LatLng
+  , H3ErrorCodes
   , H3Index
   , c2hs_cellToVertex
   , c2hs_vertexToLatLng
